+++ hutool-core/src/test/java/cn/hutool/core/date/LocalDateTimeUtilTest.java
@@ -54,6 +54,24 @@ public class LocalDateTimeUtilTest {
 	}
 
 	@Test
+	public void parseTest5() {
+		LocalDateTime localDateTime = LocalDateTimeUtil.parse("19940121183604", "yyyyMMddHHmmss");
+		Assert.assertEquals("1994-01-21T18:36:04", localDateTime.toString());
+	}
+
+	@Test
+	public void parseTest6() {
+		LocalDateTime localDateTime = LocalDateTimeUtil.parse("19940121183604682", "yyyyMMddHHmmssSSS");
+		Assert.assertEquals("1994-01-21T18:36:04.682", localDateTime.toString());
+
+		localDateTime = LocalDateTimeUtil.parse("1994012118360468", "yyyyMMddHHmmssSS");
+		Assert.assertEquals("1994-01-21T18:36:04.680", localDateTime.toString());
+
+		localDateTime = LocalDateTimeUtil.parse("199401211836046", "yyyyMMddHHmmssS");
+		Assert.assertEquals("1994-01-21T18:36:04.600", localDateTime.toString());
+	}
+
+	@Test
 	public void parseDateTest() {
 		LocalDate localDate = LocalDateTimeUtil.parseDate("2020-01-23");
 		Assert.assertEquals("2020-01-23", localDate.toString());